requirements.txt
flask==3.0.0
flask-socketio==5.3.0
opencv-python-headless==4.8.0.76
mediapipe==0.10.9
numpy==1.26.0
joblib==1.3.2
pillow==10.0.0
gunicorn==21.2.0
eventlet==0.33.3
